renamed scaleValue to adhere to Go naming convention. Did not need to be capitalized since it was not exported

+++ pkg/color/spectrum.go
@@ -16,7 +16,7 @@ func GenerateSpectrum(w, h int32) *rl.Image {
 
 			imageRGB := rl.ColorFromHSV(hue, saturation, brightness)
 			rl.ImageDrawPixel(
-				image, int32(x), int32(y), imageRGB,
+				image, x, y, imageRGB,
 			)
 		}
 	}
